refactor(api): extract module format validation
