+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/validation/element/index.js
@@ -1,0 +1,7 @@
+import { imageElementSchema } from './image.js';
+import { qcuElementSchema } from './qcu.js';
+import { qrocmElementSchema } from './qrocm.js';
+import { textElementSchema } from './text.js';
+import { videoElementSchema } from './video.js';
+
+export { imageElementSchema, qcuElementSchema, qrocmElementSchema, textElementSchema, videoElementSchema };
