+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/validation/element/video.js
@@ -1,0 +1,14 @@
+import Joi from 'joi';
+import { uuidSchema } from '../utils.js';
+
+const videoElementSchema = Joi.object({
+  id: uuidSchema,
+  type: Joi.string().valid('video').required(),
+  title: Joi.string().required(),
+  url: Joi.string().required(),
+  subtitles: Joi.string().required(),
+  transcription: Joi.string().required(),
+  alternativeText: Joi.string().required(),
+}).required();
+
+export { videoElementSchema };
