+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/validation/utils.js
@@ -1,0 +1,7 @@
+import Joi from 'joi';
+
+const uuidSchema = Joi.string().guid().required();
+
+const proposalIdSchema = Joi.string().regex(/^\d+$/);
+
+export { uuidSchema, proposalIdSchema };
